requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
openai==1.91.0
python-dotenv==1.1.0
python-multipart==0.0.20